+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Add logo in organization serializer
 - Configure language through environment variables
 - Send email when the payment is successful
 - Transform `mjml` files (email's template files) to html and plaintext